requirements.txt
feedparser>=6.0
beautifulsoup4>=4.12
lxml>=5.0
requests>=2.31
playwright>=1.40
openai>=1.12
google-auth>=2.28
tweepy>=4.14
google-api-python-client>=2.100
scikit-learn>=1.5
jsonschema>=4.0
google-cloud-discoveryengine>=0.11.0
